FIX: more reasonable mechanism to search for nodes concerning its hyperlink attribute
when trying to find fitting nodes dependent on the "hyperlink" attribute, the "exact" parameter will tolerate both, slash (e.g. Linux) or backslash (e.g. Windows) chars equally. If "exact" is not set, the search string is attempted to be found within the hyperlink string of a node, disregarding slash/backslash and uppercase/lowercase chars.