NEW: ability to choose case-sensitive or case-insensitive comparison
when setting the "exact" parameter to True, comparisons of textual content will be done regarding equality of case-insensitive content. by setting "casesensitive" to True, the comparison will be true only if this more strict condition is met. This feature can be used for the following methods: map.find_nodes(), node.find_nodes() and node.find_children(). this feature is usable in combination with the following fearch parameters: "core", "details" and "link".