MOD: searching for specific node attributes within a map or a tree now defaults to "approximate"
"approximate" search mode will search for the given string within the respective attribute value in a case-insensitive way. to set a different mode, the boolean "exact" parameter can be used (as for core, details, link, notes, ...) possibly in combination with further boolean parameters "caseinsensitive" and "generalpathsep" to disregard the case and the os-specific path separator.