NEW: node method to get the visible text content (core-linked)
using Groovy scripting or formulas, it is possible to show a distant node's content within the local node by e.g. inserting "= ID_12345678.text" into the local node. When using the "plaintext" attribute, this small formula is returned. When using the new "visibletext" attribute, the text string of the distant node (as visible within the local node) is returned